+++ app/src/main/java/com/example/pae_project/MainActivity.java
@@ -84,12 +84,6 @@ public class MainActivity extends AppCompatActivity implements PermissionsListen
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        Boolean firstRun = getSharedPreferences("PREFERENCE",MODE_PRIVATE).getBoolean("firstRun",true);
-
-        if(firstRun) {
-            Intent intro = new Intent(getApplicationContext(),Intro.class);
-            startActivity(intro);
-        }
 
         Mapbox.getInstance(this, getString(R.string.access_token));
         setContentView(R.layout.activity_main);
