+++ app/src/main/res/layout/activity_intro.xml
@@ -4,6 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Intro">
+    android:textColor="#0fff"
+    tools:context=".IntroInicial">
 
 </android.support.constraint.ConstraintLayout>
