+++ app/src/main/res/values/colors.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#008577</color>
-    <color name="colorPrimaryDark">#00574B</color>
+    <color name="colorPrimary">#00abff</color>
+    <color name="colorPrimaryDark">#006ca1</color>
     <color name="colorAccent">#D81B60</color>
 </resources>
